hypervisor: Refactor `save_pending_tables` of Vgic
